conversions and tests wip
