+++ src/services/pathContents/commit/dir/useRepoCommitContents.tsx
@@ -103,7 +103,6 @@ export const useRepoCommitContents = ({
       }).then((res) => {
         const parsedRes = RequestSchema.safeParse(res?.data)
         if (!parsedRes.success) {
-          console.log('fail', parsedRes.error)
           return Promise.reject({
             status: 404,
             data: {},
@@ -143,7 +142,7 @@ export const useRepoCommitContents = ({
         if (pathContentsType === 'PathContents') {
           results = data?.owner?.repository?.commit?.pathContents?.results
         }
-        console.log(results, 'results')
+        console.log(results)
         return {
           results: results ?? null,
           indicationRange:
